Adds fist/punch hand pose
This commit adds a new pose via cherry-picking commit [1]. The
cherry-picked commit actually contains two hand poses.

The first hand pose is a fist. In the original commit it was named
"yumruk.yml" which, for completeness, translates to 'punch' from Turkish.

The second hand pose is an open palm. In the original commit it was named
"yandan_avuc.yaml", which for completeness, translates to 'handful' from
Turkish. This pose is same as the existing relaxed.yml, and diffing the
two poses reveals the only difference is the camera angle.

Therefore only the fist/punch pose has been included. A rendered image for
the fist has been added.

(cherry picked from commit [1])
[1] https://github.com/galatasarayuniversity/libhand/commit/51e5065c5b5d0252f383cf205d357ce23d1746da