Drops OGRE 1.7.3 static installer
The existing static installer script does not succeed on any version of Ubuntu
currently supported (as of May 2016). I have succeeded in using in the past,
but I had to construct a virtual machine with an on older, unsupported Ubuntu
version (I believe 11.04 or 11.10).

The option to static linking against OGRE without much effort would still be
very useful, but the script is very brittle which failure leaves a bad first
impression for new users.