Adds LibHand/OGRE integration test
This commit adds a simple test (based on the bend_one_finger example), which
simple renders a hand, counts the number of non-black pixels. If the number is
greater than some arbitrary constant (10 was chosen), the test is considered a
success, showing that OGRE was able to load and configure its plugins and
render an image -- a basic behaviour that is surprisingly unreliable between
different configurations and platforms.

If rendered images were bit-for-bit identical, it would have been more robust
to do a file comparison as an integration test. This is unlikely between
platforms.

This patch also integrates it into Travis CI, and AppVeyor.