Updates CI status images, links to GitHub organization
Recently, the primary remote for libhand was moved from [1] to [2].

This means the Travis CI and AppVeyor status image paths in the README.md need
to be updated. Additionally, the URL was not updated to point to [1] before #4
was merged.

While Travis CI supports GitHub organisations well, AppVeyor unfortunately
required an entirely new account to be created with the account name choosing
the desired URL [3].  Thus 'libhand' was chosen as an AppVeyor account name.

[1] https://github.com/jonkeane/libhand
[2] https://github.com/libhand/libhand
[3] http://help.appveyor.com/discussions/questions/1154-appveyor-account-for-github-organizations