Added osx to travis-ci
Travis CI does not appear to have a mechanism for allowing the headless
execution to open up a virtual window. [1] The Ubuntu build uses X virtual
framebuffer, but there is no equivalent mechanism currently available on Travis
CI on macOS.

Therefore, the ogre_rendering_test only is executed on the Ubuntu build.
However, the integration test is built on all platforms.

[1] https://github.com/travis-ci/travis-ci/issues/7277